requirements.txt
Django==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-cors-headers==4.0.0
django-filter==23.3
gunicorn==20.1.0
whitenoise==6.2.0
stripe==7.8.0
requests==2.31.0
cloudinary==1.33.0
dj-database-url==2.1.0
twilio==8.10.0
django-email-verification==0.3.3
